+++ example/lib/main.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+import 'package:example/http_client_example.dart';
+
+void main() {
+  runApp(const MyApp());
+}
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Pivox Example',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+        fontFamily: 'Inter',
+      ),
+      home: const HttpClientExample(),
+    );
+  }
+}