+++ lib/src/domain/proxy_validator.dart
@@ -1,0 +1,5 @@
+import 'package:pivox/pivox.dart';
+
+abstract class ProxyValidator {
+  Future<bool> validate(Proxy proxy);
+}